refactoed dashboard to be under 25  lines. deleted Create profile component
