--- src/components/Form_Components/Create_Profile/createProfileStyle.css
@@ -1,74 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Exo&display=swap");
-.logo_name {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  margin-top: 44px;
-}
-.ProfileForm {
-  display: flex;
-  flex-direction: column;
-  margin: 20px 20px 0 20px;
-  height: fit-content;
-}
-.logo-reg {
-  width: 75px;
-  height: 50px;
-}
-.name-reg {
-  margin-top: 10px;
-  /* width: 140px; */
-  height: 30px;
-}
-.sign {
-  font-family: Exo !important;
-  font-style: normal !important;
-  font-weight: 600 !important;
-  font-size: 30px !important;
-  line-height: 38px !important;
-  color: #000000 !important;
-}
-.label {
-  text-align: left;
-  font-family: Exo !important;
-  font-style: normal !important;
-  font-weight: normal !important;
-  font-size: 14px !important;
-  line-height: 22px !important;
-  color: #262626 !important;
-}
-.labels {
-  margin-top: 30px;
-  text-align: left;
-  }
-.confirmPassword {
-  margin-top: 30px;
-  text-align: left;
-  }
-  .continueBtn{
-    margin-top: 20px !important;
-    font-family: "Exo", sans-serif !important;
-    text-transform: none !important;
-    font-size: 16px !important;
-    background: #1890ff !important ;
-    border: 1px solid #1890ff !important;
-    color: #ffffff !important;
-  }
-  .skipBtn{
-    margin-top: 20px !important;
-    font-family: "Exo", sans-serif !important;
-    text-transform: none !important;
-    font-size: 16px !important;
-    background: #ffffff !important ;
-    border: 1px solid #1890ff !important;
-    color: #1890ff !important;
-    margin-bottom: 20px !important;
-
-  }
-  .account{
-    text-align: center !important;
-    display: flex;
-    flex-direction: row;
-    font-family: "Exo", sans-serif !important;
-
-  }
