+++ src/components/Loading.js
@@ -1,0 +1,31 @@
+import React from "react";
+import { connect } from 'react-redux';
+import Loader from 'react-loader-spinner';
+
+export const Loading = props => {
+	return (
+		<div id='loading'>
+			{props.plaidFetching || props.blockFetching || props.profileFetching ? (
+				<Loader
+					type='ThreeDots'
+					color='#66aabc'
+					height={50}
+					width={50}
+					timeout={10000} //3 secs
+				/>
+			) : ( <p></p> )}
+		</div>
+	);
+};
+
+function mapStateToProps(state) {
+	return {
+		plaidFetching: state.plaidReducer.isFetching,
+		blockFetching: state.blockReducer.isFetching,
+		profileFetching: state.profileReducer.isFetching
+	};
+}
+
+export default connect(mapStateToProps, {})(
+	Loading
+);
